Longs were causing error with strategy so switched to ints

+++ YPC backend/src/main/java/com/cryptic/ypc/dal/UserRepository.java
@@ -8,13 +8,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.cryptic.ypc.model.user.User;
 
-public interface UserRepository extends JpaRepository<User, Long>{
+public interface UserRepository extends JpaRepository<User, Integer>{
 	public boolean existsById(int id);
 	
 	@Modifying
 	@Transactional
 	@Query(value = "UPDATE users SET type = 'Registered', user_name = :name, password = :password "
 			+ " WHERE user_id = :id", nativeQuery = true)
-	public void updateGuestUserToRegistered(@Param("id") Long id, @Param("name") String username,
+	public void updateGuestUserToRegistered(@Param("id") Integer id, @Param("name") String username,
 			@Param("password") String password);
 }
